+++ lambda-authorizer.tf
@@ -84,7 +84,8 @@ resource "aws_lambda_function" "authorizer" {
 
   lifecycle {
     ignore_changes = [
-      source_code_hash
+      source_code_hash,
+      filename
     ]
   }
 }