+++ lambda-stats.tf
@@ -101,6 +101,7 @@ resource "aws_lambda_function" "stats" {
   lifecycle {
     ignore_changes = [
       source_code_hash,
+      filename
     ]
   }